+++ packages/extensions/venia-payments-braintree/README.md
@@ -1,0 +1,12 @@
+# venia-payments-braintree
+
+This package provides Braintree payments extension for Venia.
+
+
+## Installation
+
+To install this pack, just add it as dependency to `venia-concept` or your
+project:
+
+*Example:*
+`yarn workspace @magento/venia-concept add @magento/venia-payments-braintree`
