+++ packages/extensions/venia-payments-braintree/package.json
@@ -1,0 +1,29 @@
+{
+  "name": "@magento/venia-payments-braintree",
+  "version": "0.0.1",
+  "publishConfig": {
+    "access": "public"
+  },
+  "description": "Provides Braintree payments extension for PWA Studio.",
+  "main": "./intercept.js",
+  "scripts": {
+    "clean": " ",
+    "test": "jest"
+  },
+  "repository": "github:magento/pwa-studio",
+  "license": "(OSL-3.0 OR AFL-3.0)",
+  "peerDependencies": {
+    "@magento/peregrine": "~9.0.0",
+    "@magento/pwa-buildpack": "~8.0.1",
+    "@magento/venia-ui": "~6.0.1",
+    "react": "~17.0.1",
+    "react-intl": "~5.7.0",
+    "react-router-dom": "~5.1.0",
+    "react-feather": "~2.0.3"
+  },
+  "pwa-studio": {
+    "targets": {
+      "intercept": "./intercept"
+    }
+  }
+}
