+++ packages/peregrine/lib/targets/__tests__/peregrine-targets.spec.js
@@ -124,7 +124,6 @@ test('exposes all hooks and targets', async () => {
           talons.CheckoutPage.ItemsReview.useItemsReview.wrapWith() wraps export "useItemsReview" from "CheckoutPage/ItemsReview/useItemsReview.js"
           talons.CheckoutPage.OrderConfirmationPage.useCreateAccount.wrapWith() wraps export "useCreateAccount" from "CheckoutPage/OrderConfirmationPage/useCreateAccount.js"
           talons.CheckoutPage.OrderConfirmationPage.useOrderConfirmationPage.wrapWith() wraps export "useOrderConfirmationPage" from "CheckoutPage/OrderConfirmationPage/useOrderConfirmationPage.js"
-          talons.CheckoutPage.PaymentInformation.useCreditCard.wrapWith() wraps export "useCreditCard" from "CheckoutPage/PaymentInformation/useCreditCard.js"
           talons.CheckoutPage.PaymentInformation.useEditModal.wrapWith() wraps export "useEditModal" from "CheckoutPage/PaymentInformation/useEditModal.js"
           talons.CheckoutPage.PaymentInformation.usePaymentInformation.wrapWith() wraps export "usePaymentInformation" from "CheckoutPage/PaymentInformation/usePaymentInformation.js"
           talons.CheckoutPage.PaymentInformation.usePaymentMethods.wrapWith() wraps export "usePaymentMethods" from "CheckoutPage/PaymentInformation/usePaymentMethods.js"
@@ -192,7 +191,6 @@ test('exposes all hooks and targets', async () => {
           talons.RootComponents.Category.useCategoryContent.wrapWith() wraps export "useCategoryContent" from "RootComponents/Category/useCategoryContent.js"
           talons.RootComponents.Category.useNoProductsFound.wrapWith() wraps export "useNoProductsFound" from "RootComponents/Category/useNoProductsFound.js"
           talons.RootComponents.Product.useProduct.wrapWith() wraps export "useProduct" from "RootComponents/Product/useProduct.js"
-          talons.SavedPaymentsPage.useCreditCard.wrapWith() wraps export "useCreditCard" from "SavedPaymentsPage/useCreditCard.js"
           talons.SavedPaymentsPage.useSavedPaymentsPage.wrapWith() wraps export "useSavedPaymentsPage" from "SavedPaymentsPage/useSavedPaymentsPage.js"
           talons.SearchBar.useAutocomplete.wrapWith() wraps export "useAutocomplete" from "SearchBar/useAutocomplete.js"
           talons.SearchBar.useSearchBar.wrapWith() wraps export "useSearchBar" from "SearchBar/useSearchBar.js"
