+++ packages/venia-concept/package.json
@@ -37,7 +37,9 @@
   },
   "homepage": "https://github.com/magento/pwa-studio/tree/master/packages/venia-concept#readme",
   "dependencies": {
-    "@magento/pwa-buildpack": "~8.0.1"
+    "@magento/pwa-buildpack": "~8.0.1",
+    "@magento/venia-sample-payments-checkmo":"link:/home/roettigl/vscode/pwa-studio/packages/extensions/venia-sample-payments-checkmo",
+    "@magento/venia-payments-braintree":"link:/home/roettigl/vscode/pwa-studio/packages/extensions/venia-payments-braintree"
   },
   "devDependencies": {
     "@adobe/apollo-link-mutation-queue": "~1.0.2",
